refactor : 지하철 역 풀네임 변환 컨버터로 변경
